requirements.txt
fastapi~=0.109.0
starlette~=0.35.1
PyYAML~=6.0.1
mysql-connector-python~=8.3.0
pydantic~=2.6.0
uvicorn~=0.27.0